Rename exception namespace
