+++ tests/Unit/Facades/Product.test.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Gildsmith\Contract\Product\ProductInterface;
-use Gildsmith\Product\Exception\MissingSoftDeletesException;
+use Gildsmith\Product\Exceptions\MissingSoftDeletesException;
 use Gildsmith\Product\Facades\ProductFacade as ProductFacadeConcrete;
 use Gildsmith\Product\Models\Blueprint;
 use Gildsmith\Product\Models\Product as ProductModel;
